feat: extract feed invitations management to a dedicated page and apply minor UI adjustments
