+++ frontend/src/app/(tasks)/feed/subscriptions/page.tsx
@@ -28,7 +28,6 @@ import '@/app/tasks/tasks.css'
 import '@/features/common/scrollbar.css'
 import { useIsMobile } from '@/features/layout/hooks/useMediaQuery'
 import { useTranslation } from '@/hooks/useTranslation'
-import { ThemeToggle } from '@/features/theme/ThemeToggle'
 import type { Subscription } from '@/types/subscription'
 
 /**
@@ -78,7 +77,7 @@ function SubscriptionsPageContent() {
   return (
     <div className="flex h-full flex-col bg-base">
       {/* Back button header */}
-      <div className="flex items-center justify-between px-4 py-3 border-b border-border bg-surface">
+      <div className="flex items-center justify-between px-4 py-3 border-b border-border bg-base">
         <div className="flex items-center gap-3">
           <Button variant="ghost" size="icon" className="h-9 w-9" onClick={handleBack}>
             <ArrowLeft className="h-5 w-5" />
@@ -94,7 +93,7 @@ function SubscriptionsPageContent() {
       </div>
 
       {/* Tab navigation */}
-      <div className="border-b border-border bg-surface/50">
+      <div className="border-b border-border bg-base">
         <Tabs
           value={activeTab}
           onValueChange={value => setActiveTab(value as SubscriptionTabValue)}
@@ -207,9 +206,7 @@ export default function SubscriptionsPage() {
             title={t('common:navigation.flow')}
             onMobileSidebarToggle={() => setIsMobileSidebarOpen(true)}
             isSidebarCollapsed={isCollapsed}
-          >
-            <ThemeToggle />
-          </TopNavigation>
+          />
 
           {/* Main content area - Subscriptions page content */}
           <div className="flex-1 overflow-hidden">
